refactor: post row, table컴포넌트 분리
